% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geobon_count.R
\name{geobon_count}
\alias{geobon_count}
\title{Get the count of GEOBON datasets}
\usage{
geobon_count()
}
\value{
Integer value, number of datasets on the GEOBON data portal.
}
\description{
Retrieves the total count of datasets on the GEOBON data portal.
}
\examples{
geobon_count()

}
