% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geobon_download.R
\name{geobon_download}
\alias{geobon_download}
\title{Download GEOBON EBV dataset from ID}
\usage{
geobon_download(id = NULL, path = getwd())
}
\arguments{
\item{id}{(integer) ID of the dataset}

\item{path}{(string) path to folder to which to download the data}
}
\value{
Function is used for its side effects
}
\description{
Downloads a specific GEOBON EBV dataset from ID.
}
\examples{
geobon_download(id = 2, path = tempdir())

}
