% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geobon_get.R
\name{geobon_get}
\alias{geobon_get}
\title{Get specific GEOBON EBV record}
\usage{
geobon_get(id = NULL, ebv_class = NULL, ebv_name = NULL)
}
\arguments{
\item{id}{(integer) ID of the dataset}

\item{ebv_class}{(character string) The EBV class of the dataset(s)}

\item{ebv_name}{(character string) The EBV name of the dataset(s)}
}
\value{
a \code{data.frame} with the record(s) matching the request.
}
\description{
Retrives a specific GEOBON EBV record. Note that only one of the three options
can be specified at a time.
}
\examples{
geobon_get(id = 50)
geobon_get(ebv_class = "Community composition")
geobon_get(ebv_name = "Habitat structure")

}
