% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geobon_list.R
\name{geobon_list}
\alias{geobon_list}
\alias{geobon_list_ebv}
\title{List all available GEOBON EBV datasets and EBV vaariable types on the portal}
\usage{
geobon_list()

geobon_list_ebv()
}
\value{
A \code{data.frame} for \code{geobon_list()} and a list for \code{geobon_list_ebv()}
}
\description{
The function \code{geobon_list} Retrieves the list of all datasets available
on the GEOBON data portal. Also, \code{geobon_list_ebv} returns all possible
values for EBV classes and names.
}
\examples{
geobon_list()
geobon_list_ebv()

}
