% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeobon-package.R
\docType{package}
\name{rgeobon}
\alias{rgeobon}
\title{rgeobon: the R interface to GEOBON's EBV open data portal API}
\description{
The goal of rgeobon is to provide an interface between R and the GEOBON open
data portal RESTfull API (please note that the portal is currently in beta).
It allows to consult the datasets available on the portal and download them.
}
